requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
bleach==1.5.0
boto==2.49.0
boto3==1.14.5
botocore==1.17.5
bounding-box==0.1.3
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.15
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.15.2
entrypoints==0.3
gast==0.4.0
gensim==3.8.3
graphviz==0.14.1
grpcio==1.31.0
h5py==2.10.0
html5lib==0.9999999
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446433964/work
ipdb==0.12.3
ipykernel==5.3.3
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi @ file:///tmp/build/80754af9/jedi_1592841914522/work
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.0
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
kiwisolver==1.1.0
lsq-ellipse==2.0.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
mpld3==0.5.1
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==6.0.3
numpy==1.18.5
olefile==0.46
opencv-python==4.3.0.36
orjson==3.2.2
overrides==2.8.0
packaging==20.4
pandas==1.0.1
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pycocotools==2.0.0
pycparser==2.19
Pygments==2.6.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
smart-open==2.0.0
tensorboard==1.8.0
tensorboardX==2.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==0.4.1.post2
torchvision==0.2.0
tornado==6.0.4
tqdm==4.42.1
traitlets==4.3.3
trash-cli==0.17.1.14
urllib3==1.25.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
